Move osr related files to atom/browser/osr
