+++ atom/browser/api/atom_api_web_contents.h
@@ -12,13 +12,12 @@
 #include "atom/browser/api/save_page_handler.h"
 #include "atom/browser/api/trackable_object.h"
 #include "atom/browser/common_web_contents_delegate.h"
+#include "atom/browser/osr/osr_output_device.h"
 #include "content/public/browser/web_contents_observer.h"
 #include "content/public/common/favicon_url.h"
 #include "content/common/cursors/webcursor.h"
 #include "native_mate/handle.h"
 #include "ui/gfx/image/image.h"
-
-#include "atom/browser/osr_output_device.h"
 
 namespace blink {
 struct WebDeviceEmulationParams;
