+++ atom/browser/osr/osr_output_device.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by the MIT license that can be
 // found in the LICENSE file.
 
-#include "atom/browser/osr_output_device.h"
+#include "atom/browser/osr/osr_output_device.h"
 
 #include "third_party/skia/include/core/SkDevice.h"
 #include "ui/gfx/skia_util.h"
