+++ atom/browser/osr/osr_render_widget_host_view.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by the MIT license that can be
 // found in the LICENSE file.
 
-#include "atom/browser/osr_render_widget_host_view.h"
+#include "atom/browser/osr/osr_render_widget_host_view.h"
 
 #include <vector>
