+++ filenames.gypi
@@ -206,13 +206,13 @@
       'atom/browser/native_window_mac.h',
       'atom/browser/native_window_mac.mm',
       'atom/browser/native_window_observer.h',
-      'atom/browser/osr_web_contents_view.cc',
-      'atom/browser/osr_web_contents_view.h',
-      'atom/browser/osr_output_device.cc',
-      'atom/browser/osr_output_device.h',
-      'atom/browser/osr_render_widget_host_view.cc',
-      'atom/browser/osr_render_widget_host_view.h',
-      'atom/browser/osr_render_widget_host_view_mac.mm',
+      'atom/browser/osr/osr_web_contents_view.cc',
+      'atom/browser/osr/osr_web_contents_view.h',
+      'atom/browser/osr/osr_output_device.cc',
+      'atom/browser/osr/osr_output_device.h',
+      'atom/browser/osr/osr_render_widget_host_view.cc',
+      'atom/browser/osr/osr_render_widget_host_view.h',
+      'atom/browser/osr/osr_render_widget_host_view_mac.mm',
       'atom/browser/net/asar/asar_protocol_handler.cc',
       'atom/browser/net/asar/asar_protocol_handler.h',
       'atom/browser/net/asar/url_request_asar_job.cc',
